Move ZIP building to separate classes
