+++ webscrambles/src/main/kotlin/org/worldcubeassociation/tnoodle/server/webscrambles/zip/model/File.kt
@@ -1,4 +1,4 @@
-package org.worldcubeassociation.tnoodle.server.webscrambles.zip
+package org.worldcubeassociation.tnoodle.server.webscrambles.zip.model
 
 data class File(val name: String, val content: ByteArray, val parent: Folder? = null) : ZipNode {
     constructor(name: String, content: String, parent: Folder? = null) : this(name, content.toByteArray(), parent)
